+++ test/fixtures/wrapper_fixtures/port_discovery_replacement.mock.js
@@ -1,0 +1,18 @@
+
+
+require('../../..').init({
+  new_port : 20000
+});
+
+var express = require('express');
+var app = express();
+
+app.get('/', function(req, res) {
+  res.send({success:true});
+});
+
+app.get('/nothing', function(req, res) {
+  res.send('yes');
+});
+
+app.listen(9007);